split parallel_for and _scan into their own headers similar to _reduce
